#4 feat: 가게등록 페이지 contents 분리
